fix test_compare

+++ test/test_compare.py
@@ -28,7 +28,7 @@ import os
 import unittest
 from testUtils import ProcessTestCase
 
-@unittest.skipIf(os.getenv("STOREMODE")=="hdf5","HDF5 version is not implemented for this test")
+# @unittest.skipIf(os.getenv("STOREMODE")=="hdf5","HDF5 version is not implemented for this test")
 class TestCompare(ProcessTestCase):
     def setUp(self):
         'Create a project'
@@ -145,10 +145,13 @@ class TestCompare(ProcessTestCase):
     def testSimpleProj(self):
         'Test compare by site'
         # do not use the default project
-        self.runCmd('vtools init test -f')
+        self.runCmd('vtools init test -f --store '+self.storeMode)
         self.runCmd('vtools import vcf/compare.vcf --build hg18')
-        self.runCmd('vtools execute "DELETE FROM genotype.genotype_1 WHERE GT=0"')
-        self.runCmd('vtools execute "DELETE FROM genotype.genotype_2 WHERE GT=0"')
+        if self.storeMode=="sqlite":
+            self.runCmd('vtools execute "DELETE FROM genotype.genotype_1 WHERE GT=0"')
+            self.runCmd('vtools execute "DELETE FROM genotype.genotype_2 WHERE GT=0"')
+        elif self.storeMode=="hdf5":
+            self.runCmd('vtools remove genotypes "GT_geno is nan"')
         self.runCmd('vtools select variant "variant_id in (1, 2)" -t T1')
         self.runCmd('vtools select variant "variant_id in (1, 3, 4)" -t T2')
         # variant
